chore(textobjects): split up into submodules
